@@ -6,9 +6,10 @@
   "repository": "git@gitlab.kilic.dev:config/eslint-config.git",
   "author": "Cenk Kilic",
   "license": "MIT",
+  "type": "module",
   "scripts": {
-    "format": "prettier --loglevel warn --write ./*.js && eslint --ext .ts,.js,.tsx,.jsx --fix ./",
-    "lint": "eslint --ext .ts,.js,.tsx,.jsx ./"
+    "format": "prettier --log-level warn --write ./*.js && eslint --fix",
+    "lint": "eslint"
   },
   "simple-git-hooks": {
     "pre-commit": "pnpm exec lint-staged",
@@ -60,7 +61,7 @@
   },
   "devDependencies": {
     "@cenk1cenk2/cz-cc": "^1.7.6",
-    "eslint": "^8.57.0",
+    "eslint": "^9.1.1",
     "lint-staged": "^15.2.2",
     "prettier": "^3.2.5",
     "simple-git-hooks": "^2.11.1"
