@@ -1,5 +1,5 @@
 /** @type {import("prettier").Config} */
-module.exports = {
+export default {
   printWidth: 179,
   tabWidth: 2,
   useTabs: false,
