@@ -1,3 +1,1 @@
-const ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS = [ '*.js', '*.cjs', '*.mjs', '*.ts', '*.tsx', '*.vue', '*.svelte' ]
-
-module.exports = { ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS }
+export const ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS = ['*.js', '*.cjs', '*.mjs', '*.ts', '*.tsx', '*.vue', '*.svelte']
