@@ -5,7 +5,7 @@ module.exports = {
   overrides: [
     {
       files: ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS,
-      plugins: [ 'import' ],
+      plugins: ['import'],
       rules: {
         'import/no-extraneous-dependencies': 'error',
         'import/exports-last': 'error'
