@@ -1,12 +1,12 @@
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: [ 'plugin:svelte/recommended', './typescript-dynamic' ],
+  extends: ['plugin:svelte/recommended', './typescript-dynamic'],
   parserOptions: {
-    extraFileExtensions: [ '.svelte' ]
+    extraFileExtensions: ['.svelte']
   },
   overrides: [
     {
-      files: [ '*.svelte' ],
+      files: ['*.svelte'],
       parser: 'svelte-eslint-parser',
       parserOptions: {
         parser: '@typescript-eslint/parser'
