@@ -1,20 +1,20 @@
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: [ './typescript' ],
+  extends: ['./typescript'],
   overrides: [
     {
-      files: [ '*.ts', '*.tsx' ],
+      files: ['*.ts', '*.tsx'],
       parserOptions: {
         project: 'tsconfig.json'
       },
       rules: {
         '@typescript-eslint/no-floating-promises': 'error',
-        '@typescript-eslint/return-await': [ 'error', 'never' ],
+        '@typescript-eslint/return-await': ['error', 'never'],
         '@typescript-eslint/await-thenable': 'error',
         'dot-notation': 'off',
-        '@typescript-eslint/dot-notation': [ 'error' ],
+        '@typescript-eslint/dot-notation': ['error'],
         'no-throw-literal': 'off',
-        '@typescript-eslint/no-throw-literal': [ 'error' ],
+        '@typescript-eslint/no-throw-literal': ['error'],
         '@typescript-eslint/prefer-reduce-type-parameter': 'error',
         '@typescript-eslint/consistent-type-imports': [
           'error',
@@ -23,7 +23,7 @@ module.exports = {
             disallowTypeAnnotations: true
           }
         ],
-        '@typescript-eslint/consistent-type-exports': [ 'error', { fixMixedExportsWithInlineTypeSpecifier: false } ]
+        '@typescript-eslint/consistent-type-exports': ['error', { fixMixedExportsWithInlineTypeSpecifier: false }]
       }
     }
   ]
