@@ -1,17 +1,17 @@
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: [ './index' ],
+  extends: ['./index'],
   overrides: [
     {
-      files: [ '*.vue' ],
-      extends: [ '@nuxtjs' ],
+      files: ['*.vue'],
+      extends: ['@nuxtjs'],
       rules: {
         'vue/singleline-html-element-content-newline': 0,
         'vue/multiline-html-element-content-newline': 0,
         'vue/html-self-closing': 0,
         'vue/no-v-html': 0,
-        'array-element-newline': [ 'error', 'consistent' ],
-        'array-bracket-newline': [ 'error', 'consistent' ],
+        'array-element-newline': ['error', 'consistent'],
+        'array-bracket-newline': ['error', 'consistent'],
         'array-bracket-spacing': [
           'error',
           'always',
